Created rexec.pl
rexec.pl is a remote execute script. This one is oriented towards single
execution though it can handle multiple machines and multiple commands
it's sorta klunky. It also does not have any of the parallelization and
logging facilities that were envisoned for rexec. The script rexec still
exists and should be integrated into rexec.pl over time. Still need to
make a Machines.pm modules to link to.