Added Subject to email address
This is not perfect as we reach into the @msgs array to grep the first
subject line. A sender with multiple messages may very well have
different subject lines. Also we needed to make a copy of @msgs.
Apparently a reference into the @msgs array messes up the later pop
operation.