Fixed announceEmail.pl to be less busy
Turns out announceEmail.pl was looping in MonitorMail. The right way to
do this is to call $IMAP->idle to set the call back. After that
announceEmail.pl is suspended in idle. When action happens on the IMAP
server the callback is called. So MonitorMail gets run to process the
callback. But then the callback is gone and needs to be re-established.