Changes are:
. Added another greeting to announceEmail.pl
. Removed a Perl::Critic warning for Unils.pm
. Changed message for Speak to use [silent]
. Created announceEmail sleep script

Actually not a sleep script but rather it is a script that needs to be
added to /lib/systemd/systemd-sleep (symlinked) so that when Ubuntu
20.04 comes out of suspend mode a USR2 signal is sent to all
announceEmail.pl processes. They then wake up and reconnect to their
IMAP hosts. This is necessary because the network is down during suspend
and Mail::IMAPTalk doesn't notice this so email announcements stop
happening. If instead Mail::IMAPTalk would have noticed this and
returned an error I could re-establish the connection in
announceEmail.pl but it doesn't.

Notice also that it takes 10-30 seconds to re-establish the network
connection after suspend so we just put a 60 second sleep in this script
rather than implementing a more complicated polling loop looking for the
network to come back up.