Fixed long standing bug about displaying proper message
If there were more than one message sent by the sender then attempting
to display the individual message would not work. This was because we
were going by msg_nbr but the msg_nbr did not really relate to the
message shown on the detail.cti web page. Here we change the alogrithm
to use the message date instead as it uniquely identifies the message.

Note that there still may be times when one sender/msg_date relates to
more than one message (think two messages with the exact same
timestamp). I'm not sure if this is possible nor if it's really that
important.