Certbot renewal hook scripts
These scripts help automate Let's Encrypt domain renewal. The certbot
renew command offers --manual-authenticate-hook and
--manual-cleanup-hook to automate the process of renewal. For DNS domain
validation the authenticate script will be called with CERTBOT_DOMAIN
and CERTBOT_VALIDATION values filled in. We then add a TXT record to the
domain with the supplied value and enter into a loop checking for
propagation. If propagated the hook returns to certbot which renews the
cert. Cleanup consists of removing the TXT record.