Rewrote ReadMsg - Andrew@ClearSCM.com check
As I've gotten Andrew@ClearSCM.com email working again and it just send
it to Andrew@DeFaria.com I noticed that I was now getting additional
spam messages with a To of Andrew@ClearSCM.com. I rarely, if ever send
email to Andrew@ClearSCM.com and if I start I'll use the old trick of a
checking for full email address of "Andrew DeFaria
<Andrew@ClearSCM.com>" on To (not implemented yet). Meantime I will just
assume all email sent to Andrew@ClearSCM.com is spam.

In order to implement this I needed to change ReadMsg to also return the
To: address. Since the return arguments were getting large I chqanged
this to return a hash.