Hopefully final adjustments to certbot
A few problems with this latest cert renewal. While the renewal happened it put
the certs in /System/Data/Certificates whereas other references were to
/System/Certificates. This has been fixed.

Additionally, certbot_deploy.sh was never called! Added it to
certbot_cleanup.sh.

An additional problem was that the crontab entry had "*/3" for month and while
that says "every 3 months" it doesn't say which month it starts on. So I changed
this to explicitedly state "Jan,Apr,Jul,Oct".

Finally, the cert generated on 12/23/2022 said it was valid until 01/22/2023.
Not sure why this was one month so we may have a failure in certs come February.
If so we'll need to check why were are only getting 30 day certs issued by LetsEncrypt.